Merge in LIVTRAVEL/order-flight from release/developer to master

* commit 'eddfcbbd904915b88cf968e8c6bf4e7c01b3fc4e': (63 commits)
  Ajustes testes unitários #deploydev
  Ajustes ordenação items request #deploydev
  Ajustes testes unitários #deploydev
  Ajustes testes integrados: - Setando corretamente valores do flight e tax - Ajustando busca sempre pelo commerceOrderId - Adicionando novos itens de multiplicadores e markup vindos da calculadora
  Ajustes testes unitários
  Ajustes testes integrados
  Ajuste ordenação entidades
  refactor ReservationServiceImpl #deploydev
  Ajustes fluxo reservation
  Ajustes erros priceModality
  Ajuste no reservation response #deploydev
  definindo fk como nullable #deploydev
  executando migration atualizada #deploydev
  executando migration atualizada
  executando migration
  executando migration
  executando migration
  revert
  Criando nova migration
  desfazendo migration
  ...

+++ src/main/java/br/com/livelo/orderflight/domain/dtos/pricing/response/PricingCalculatePrice.java
@@ -12,9 +12,9 @@ import java.math.BigDecimal;
 public class PricingCalculatePrice {
     private String priceListId;
     private BigDecimal amount;
-    private Integer pointsAmount;
+    private BigDecimal pointsAmount;
     private PricingCalculateFlight flight;
     private PricingCalculateTaxes taxes;
-    private Integer accrualPoints;
+    private BigDecimal accrualPoints;
     private PricingCalculatePricesDescription pricesDescription;
 }
